+++ src/sections/traits/_traits.js
@@ -1,6 +1,8 @@
 import animation from "./animation.js";
 import audio from "./audio.js";
+import elevation from "./elevation.js";
 import files from "./files.js";
+import levels from "./levels.js";
 import moves from "./moves.js";
 import opacity from "./opacity.js";
 import rotation from "./rotation.js";
@@ -18,7 +20,9 @@ import name from "./name.js";
 export default {
   animation,
   audio,
+  elevation,
   files,
+  levels,
   moves,
   opacity,
   rotation,